% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_specific_fdr.R
\name{feature_fdr_correction}
\alias{feature_fdr_correction}
\title{feature_fdr_correction}
\usage{
feature_fdr_correction(toptable, interesting_features_table, method = "BH")
}
\arguments{
\item{toptable}{A data frame with the results of the limma analysis as obtained from the topTable function. It should contain the columns 'P.Value' and 'adj.P.Val' and 'nterm_modif_peptide'.}

\item{interesting_features_table}{A subset of the peptide data annotation data frame obtained from the \code{annotate_neo_termini} function. It should contain the list of features that can be defined as 'neo-termini' under the user criteria. Example: All peptides with TMT modification at the N-termini.}

\item{method}{The method to be used for multiple testing correction. Default is "BH" (Benjamini-Hochberg).}
}
\description{
Apply multiple-testing p-value adjustment on specified features.
This approach is based on the idea of independent hypothesis weighting,
where the FDR is calculated for a subset of interesting features independently of the tested hypothesis of differential abundance.
As described by https://doi.org/10.1038/nmeth.3885
}
\details{
Apply multiple-testing p-value adjustment on specified features
}
\author{
Miguel Cosenza-Contreras
}
