% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{fragpipe_adapter}
\alias{fragpipe_adapter}
\title{fragpipe_adapter}
\usage{
fragpipe_adapter(
  parent_dir,
  ref_sample = "norm",
  grouping_var = "nterm_modif_peptide",
  min_purity = 0.5,
  tmt_delta = "229"
)
}
\arguments{
\item{parent_dir}{directory of FragPipe search results}

\item{ref_sample}{define the reference channel in multi-mixture TMT experiments. Default is NULL (1 mixture). Virtual reference is not available.}

\item{grouping_var}{define the grouping variable for MAD scaling. Default is "nterm_modif_peptide" for normalization at the feature level of "Peptide + N-terminal modification".}

\item{min_purity}{minimum purity of the PSMs to be considered. Default is 0.5.}

\item{tmt_delta}{define the TMT reporter ion mass delta. Default is "229" for TMT10/11plex. "304" for TMT16plex.}
}
\value{
A data frame with at least 4 columns
\describe{
\item{nterm_modif_peptide}{Peptide identification merging N-terminal modification + peptide sequence}
\item{nterm_modif}{N-terminal modification}
\item{peptide}{Peptide sequence}
\item{protein}{Protein ID based on Uniprot Accession}
}
}
\description{
Process psm.tsv files from FragPipe search results for further analysis with TermineR
}
\details{
Process psm.tsv files from FragPipe search results for further analysis with TermineR
}
\author{
Miguel Cosenza-Contreras
}
