% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{fragpipe_dda_heavy_light_adapter}
\alias{fragpipe_dda_heavy_light_adapter}
\title{fragpipe_dda_heavy_light_adapter}
\format{
A data frame with at least 4 columns
\describe{
\item{nterm_modif_peptide}{Peptide identification merging N-terminal modification + peptide sequence}
\item{nterm_modif}{N-terminal modification}
\item{peptide}{Peptide sequence}
\item{protein}{Protein ID based on Uniprot Accession}
}
}
\usage{
fragpipe_dda_heavy_light_adapter(
  combined_modified_peptide_file,
  annotation_file_path,
  grouping_var = "nterm_modif_peptide"
)
}
\arguments{
\item{combined_modified_peptide_file}{path to file with FragPipe combined_modified_peptide_label_quant.tsv file}

\item{annotation_file_path}{path to the annotation file with sample names. It should contain at least two colums: \code{run}, matching the name of the LC-MS/MS run, and \code{sample_name}, with a readable sample identifier.}

\item{grouping_var}{define the grouping variable for MAD scaling. Default is "nterm_modif_peptide" for normalization at the feature level of "Peptide + N-terminal modification".}
}
\description{
Process combined_modified_peptide_label_quant.tsv file from FragPipe label-free search results of isotopic labelling on MS1 level, for further analysis with TermineR
}
\details{
Process combined_modified_peptide_label_quant.tsv file from FragPipe label-free search results of isotopic labelling on MS1 level, for further analysis with TermineR
}
\author{
Miguel Cosenza-Contreras
}
