% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{fragpipe_lf_adapter}
\alias{fragpipe_lf_adapter}
\title{fragpipe_lf_adapter}
\format{
A data frame with at least 4 columns
\describe{
\item{nterm_modif_peptide}{Peptide identification merging N-terminal modification + peptide sequence}
\item{nterm_modif}{N-terminal modification}
\item{peptide}{Peptide sequence}
\item{protein}{Protein ID based on Uniprot Accession}
}
}
\usage{
fragpipe_lf_adapter(parent_dir, annotation_file_path, grouping_var)
}
\arguments{
\item{parent_dir}{directory of FragPipe search results}

\item{annotation_file_path}{path to the annotation file with sample names. It should contain at least two colums: \code{run}, matching the name of the LC-MS/MS run, and \code{sample_name}, with a readable sample identifier.}

\item{grouping_var}{define the grouping variable for MAD scaling. Default is "nterm_modif_peptide" for normalization at the feature level of "Peptide + N-terminal modification".}
}
\description{
Process psm.tsv files from FragPipe search results for further analysis with TermineR
}
\details{
Process psm.tsv files from FragPipe label-free search results for further analysis with TermineR
}
\author{
Miguel Cosenza-Contreras
}
