% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_and_isoforms_uniprot_processing}
\alias{human_and_isoforms_uniprot_processing}
\title{Human protein processing annotation from Uniprot API. Swissprot including isoforms.}
\format{
A data frame
\describe{
\item{type}{Type of processing feature. One of SIGNAL, CHAIN, INIT_MET, TRANSIT, PEPTIDE}
\item{description}{Description of processing feature}
\item{begin}{begin of the processing feature in protein sequence}
\item{end}{end of the processing feature in protein sequence}
\item{length}{length of the processing feature within protein sequence}
\item{accession}{Uniprot accession ID associated with the feature}
\item{entryName}{Uniprot Entry name ID associated with the feature}
\item{taxid}{Organism tax id}
\item{order}{order}
}
}
\source{
Uniprot API (queried on August 2024)
}
\usage{
human_and_isoforms_uniprot_processing
}
\description{
Tabular annotation of processing information as obtained from Uniprot API
}
\keyword{datasets}
