% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_iso_targetp_processing}
\alias{human_iso_targetp_processing}
\title{Human TargetP processing annotation. Swissprot including isoforms.}
\format{
A data frame
\describe{
\item{protein}{Uniprot accession ID}
\item{targetp_category}{TargetP prediction category}
\item{targetp_p1_position}{TargetP predicted P1 cleavage position}
}
}
\source{
TargetP 2.0 prediction output
}
\usage{
human_iso_targetp_processing
}
\description{
Tabular annotation of TargetP-predicted processing sites
}
\keyword{datasets}
