% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{merops_accession_to_protease}
\alias{merops_accession_to_protease}
\title{MEROPS protease ID to name mapping}
\format{
A data frame with:
\describe{
\item{merops_protease_id}{MEROPS protease identifier (e.g., C14.003)}
\item{protease_name}{Protease name}
}
}
\source{
MEROPS database, current_release/database_files/uniprot.txt;
ftp://ftp.ebi.ac.uk/pub/databases/merops/current_release/database_files/ (accessed September 2025).
}
\usage{
merops_accession_to_protease
}
\description{
Two-column lookup mapping MEROPS internal protease identifiers to human-readable protease names.
}
\details{
Derived from the MEROPS uniprot.txt mapping file and deduplicated.
}
\keyword{datasets}
