% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermineR-package.R
\docType{package}
\name{TermineR-package}
\alias{TermineR}
\alias{TermineR-package}
\title{TermineR: Annotation and processing of shotgun proteomics semi-specific search results for the analysis of proteolytic processing (terminomics analysis).}
\description{
Data analysis approach for the annotation, and differential analysis of proteolytic products (i.e. semi-specific peptides) from shotgun proteomics search results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MiguelCos/TermineR}
  \item Report bugs at \url{https://github.com/MiguelCos/TermineR/issues}
}

}
\author{
\strong{Maintainer}: Miguel Cosenza-Contreras \email{migueljcc5@gmail.com}

}
\keyword{internal}
