% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unimod_id_to_name_mapping}
\alias{unimod_id_to_name_mapping}
\title{UniMod ID to modification name mapping}
\format{
A data frame
\describe{
\item{id}{UniMod ID}
\item{id_nr}{Numerical UniMod ID}
\item{name}{Descriptive name of the modification}
\item{monoisotopic_mass}{Monoisotopic mass delta of the modification}
\item{average_mass}{Average mass delta of the modification}
\item{composition}{Chemical composition delta of the modification}
}
}
\source{
UniMod OBO (queried on June 2026)
}
\usage{
unimod_id_to_name_mapping
}
\description{
Tabular annotation of UniMod IDs, names, masses, and compositions
}
\keyword{datasets}
