% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_neo_termini.R
\name{annotate_neo_termini}
\alias{annotate_neo_termini}
\title{annotate_neo_termini}
\format{
A data frame with at least 28 columns
\describe{
\item{nterm_modif_peptide}{Peptide identification merging N-terminal modification + peptide sequence}
\item{nterm_modif}{N-terminal modification}
\item{peptide}{Peptide sequence}
\item{protein}{Protein ID based on Uniprot Accession}
\item{protein_length}{Protein length}
\item{peptide_start}{Start position of the peptide in the protein sequence}
\item{peptide_end}{End position of the peptide in the protein sequence}
\item{last_aa}{Last amino acid of the peptide}
\item{first_aa}{First amino acid of the peptide}
\item{aa_after}{Amino acid after the peptide}
\item{aa_before}{Amino acid before the peptide}
\item{sense}{Direction of the peptide cleavage ("C" means after C-termini, like trypsin)}
\item{specificity}{Specificity of the cleavage sites of the peptide ("specific", "semi_Nterm", or "semi_Cterm")}
\item{five_res_before}{Five amino acids before the cleavage site}
\item{five_res_after}{Five amino acids after the cleavage site}
\item{cleavage_site}{Five amino acids before and after the cleavage site, separated by "|"}
\item{cleavage_sequence}{Ten amino acids around the cleavage site}
\item{p1_position}{Position of the P1 residue around the cleavage site}
\item{p1_prime_position}{Position of the P1' residue around the cleavage site}
\item{p1_residue}{P1 residue (amino acid)}
\item{p1_prime_residue}{P1' residue (amino acid)}
\item{p1_position_percentage}{Position of the P1 residue in proportion of the protein length}
\item{met_clipping}{Logical. TRUE if the peptide is a methionine clipping}
\item{matches_p1_prime}{Logical. TRUE if the peptide matches the P1' position of a processing feature}
\item{uniprot_processing_type}{Type of processing feature from Uniprot API. One of "INIT_MET", "PROPEP",
"SIGNAL", "TRANSIT", "CHAIN".
"non_canonical" indicates that the potential cleavage location
has not been annotated in UniProt processing features.
"not_canonical_no_procc_annot" indicates that the potential cleavage site
maps to a Uniprot protein ID that does not have processing features information.}
\item{processing_annotation_start}{Start position of the processing feature in the protein sequence}
\item{processing_annotation_end}{End position of the processing feature in the protein sequence}
\item{protein_sequence}{Protein sequence}
\item{sample_columns}{Scaled abundances of annotated features per sample}
}
}
\usage{
annotate_neo_termini(
  peptides_df,
  fasta_location,
  sense,
  specificity,
  organism,
  n_residues_area = 10,
  distinct = TRUE
)
}
\arguments{
\item{peptides_df}{peptide data frame as obtained from one of the adapter functions. Must contain columns nterm_modif_peptide, nterm_modif, peptide, protein and one quantitative column per sample.}

\item{fasta_location}{location of the fasta file containing the protein sequences identified in the sample, in uniprot format.}

\item{sense}{direction of the peptide cleavage. One of "N" or "C".}

\item{specificity}{amino acid specificity of the cleavage site. Example: "K|R" for trypsin.}

\item{organism}{organism of the fasta file. Default is "mouse". Other options are "human", "medicago_trucantula", "rhizobium_melitoli", "pig", "human_iso", "ecoli", "arabidopsis", "rat", "yeast", and "c_elegans".}

\item{n_residues_area}{number of residues to consider around the cleavage area. Default is 10. Example: if n_residues_area = 10, the cleavagea are would be 10 residues before and 10 residues after the cleavage site.}

\item{distinct}{logical. If TRUE, keep only one peptide sequence per feature. Default is TRUE.}
}
\description{
Annotate peptides and associated cleavage events.
}
\details{
Annotate peptides and associated cleavage events
}
\author{
Miguel Cosenza-Contreras
}
