% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{diann_adapter}
\alias{diann_adapter}
\title{diann_adapter}
\format{
A data frame with at least 4 columns
\describe{
\item{nterm_modif_peptide}{Peptide identification merging N-terminal modification + peptide sequence}
\item{nterm_modif}{N-terminal modification. Currently no N-terminal modification is supported for DIANN data}
\item{peptide}{Peptide sequence}
\item{protein}{Protein ID based on Uniprot Accession}
}
}
\usage{
diann_adapter(path_to_file, proteotypic = TRUE, summarization = "SUM")
}
\arguments{
\item{path_to_file}{path to report file with DIANN search results in tsv format}

\item{proteotypic}{keep only proteotypic peptides. Default is TRUE.}

\item{summarization}{quantitative summarization approach for features identified by several precursors.
"SUM" will sum up the intensities of the precursors. "MAX" will keep the intensity of the precursor with higher intensity.
Default is "SUM".}
}
\description{
Process report file from DIANN search results for further analysis with TermineR
}
\details{
Process report file from DIANN search results for further analysis with TermineR
}
\author{
Miguel Cosenza-Contreras
}
