% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_to_protein_normalization.R
\name{peptide2protein_normalization}
\alias{peptide2protein_normalization}
\title{peptide2protein_normalization}
\usage{
peptide2protein_normalization(
  peptides,
  annot,
  peptide_annot,
  summarize_by_specificity = TRUE
)
}
\arguments{
\item{peptides}{A data frame with the raw (not log2 transformed) peptide abundances per sample. It should contain the columns 'nterm_modif_peptide', 'protein' and the sample names as columns.}

\item{annot}{A data frame with the annotation of the samples. It should contain at least a 'sample' column, matching the column names for the samples in the peptide data frame.}

\item{peptide_annot}{A data frame with the annotation of the peptides. It should contain at least a 'nterm_modif_peptide' column, matching the 'nterm_modif_peptide' column in the peptide data frame, and the 'specificity' column.}

\item{summarize_by_specificity}{A logical value indicating if the protein abundances should be summarized based on specific peptides. Default is TRUE.}
}
\description{
This is an experimental function for the normalization of peptides abundances against the abundance of their respective proteins.
The protein abundances are summarized from the peptide feature abundance information based on fully specific peptides, as a proxy for the protein abundance disregarding the effect of proteolytic activity.
Then, peptide-to-protein abundance ratios are calculated based on raw intensities and log2 transformed for downstream processing.
}
\details{
Normalize peptide abundances against their respective protein abundances.
}
\author{
Miguel Cosenza-Contreras
}
