% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unimod_id_to_name_mapping}
\alias{unimod_id_to_name_mapping}
\title{UniMOd ID to modification name mapping}
\format{
A data frame
\describe{
\item{id}{UniMod ID}
\item{id_nr}{Numerical UniMod ID}
\item{name}{Descriptive name of the modification}
}
}
\source{
Uniprot API (queried on August 2024)
}
\usage{
unimod_id_to_name_mapping
}
\description{
Tabular annotation of the numerical ID from UniMod and their associated modifications
}
\keyword{datasets}
